int/disk/memory: organizes the package.
